+++ INTERX/gateway/interx/nodelist.go
@@ -1,0 +1,51 @@
+package interx
+
+import (
+	"net/http"
+
+	"github.com/KiraCore/sekai/INTERX/common"
+	"github.com/KiraCore/sekai/INTERX/config"
+	"github.com/gorilla/mux"
+	"github.com/grpc-ecosystem/grpc-gateway/runtime"
+)
+
+// RegisterInterxQueryRoutes registers query routers.
+func RegisterNodeListQueryRoutes(r *mux.Router, gwCosmosmux *runtime.ServeMux, rpcAddr string) {
+	r.HandleFunc(config.QueryNodeList, QueryNodeList(gwCosmosmux, rpcAddr)).Methods("GET")
+
+	common.AddRPCMethod("GET", config.QueryNodeList, "This is an API to query node list.", true)
+}
+
+func queryNodeListHandle(gwCosmosmux *runtime.ServeMux, rpcAddr string) (interface{}, interface{}, int) {
+	return nil, nil, http.StatusOK
+}
+
+// QueryNodeList is a function to query node list.
+func QueryNodeList(gwCosmosmux *runtime.ServeMux, rpcAddr string) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		request := common.GetInterxRequest(r)
+		response := common.GetResponseFormat(request, rpcAddr)
+		statusCode := http.StatusOK
+
+		common.GetLogger().Info("[query-node-list] Entering status query")
+
+		if !common.RPCMethods["GET"][config.QueryNodeList].Enabled {
+			response.Response, response.Error, statusCode = common.ServeError(0, "", "API disabled", http.StatusForbidden)
+		} else {
+			if common.RPCMethods["GET"][config.QueryNodeList].CachingEnabled {
+				found, cacheResponse, cacheError, cacheStatus := common.SearchCache(request, response)
+				if found {
+					response.Response, response.Error, statusCode = cacheResponse, cacheError, cacheStatus
+					common.WrapResponse(w, request, *response, statusCode, false)
+
+					common.GetLogger().Info("[query-node-list] Returning from the cache")
+					return
+				}
+			}
+
+			response.Response, response.Error, statusCode = queryNodeListHandle(gwCosmosmux, rpcAddr)
+		}
+
+		common.WrapResponse(w, request, *response, statusCode, common.RPCMethods["GET"][config.QueryStatus].CachingEnabled)
+	}
+}