+++ INTERX/main.go
@@ -70,7 +70,11 @@ func main() {
 	initFeeAmounts := initCommand.String("fee_amounts", "stake 1000ukex,ukex 1000ukex,validatortoken 1000ukex", "The fee amount for each denom. `stake 1000ukex` means it will use `1000ukex` for `stake` assets transfer.")
 
 	initAddrBook := initCommand.String("addrbook", "addrbook.json", "The address books")
+	initNodeKey := initCommand.String("node_key", "node_key.json", "The node key file path")
 	initTxModes := initCommand.String("tx_modes", "sync,async,block", "The allowed transaction modes")
+
+	initNodeDiscoveryUseHttps := initCommand.Bool("node_discovery_use_https", false, "The option to use https in node discovery")
+	initNodeDiscoveryPort := initCommand.String("node_discovery_port", "11000", "The default interx port to be used in ndoe discovery")
 
 	startConfigPtr := startCommand.String("config", "./config.json", "The interx configurtion path. (Required)")
 
@@ -117,6 +121,9 @@ func main() {
 					*initFeeAmounts,
 					*initAddrBook,
 					*initTxModes,
+					*initNodeDiscoveryUseHttps,
+					*initNodeDiscoveryPort,
+					*initNodeKey,
 				)
 
 				fmt.Printf("Created interx configuration file: %s\n", *initConfigFilePtr)