+++ x/gov/client/cli/query.go
@@ -19,7 +19,8 @@ import (
 
 // Proposal flags
 const (
-	flagVoter = "voter"
+	flagVoter   = "voter"
+	flagReverse = "reverse"
 )
 
 // GetCmdQueryPermissions the query delegation command.
@@ -262,6 +263,7 @@ $ %s query gov proposals --voter cosmos1skjwj5whet0lpe65qaq4rpq03hjxlwd9nf39lk
 		),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			bechVoterAddr, _ := cmd.Flags().GetString(flagVoter)
+			reverse, _ := cmd.Flags().GetBool(flagReverse)
 
 			if len(bechVoterAddr) != 0 {
 				_, err := sdk.AccAddressFromBech32(bechVoterAddr)
@@ -273,10 +275,17 @@ $ %s query gov proposals --voter cosmos1skjwj5whet0lpe65qaq4rpq03hjxlwd9nf39lk
 			clientCtx := client.GetClientContextFromCmd(cmd)
 			queryClient := types.NewQueryClient(clientCtx)
 
+			pageReq, err := client.ReadPageRequest(cmd.Flags())
+			if err != nil {
+				return err
+			}
+
 			res, err := queryClient.Proposals(
 				context.Background(),
 				&types.QueryProposalsRequest{
-					Voter: bechVoterAddr,
+					Voter:      bechVoterAddr,
+					Pagination: pageReq,
+					Reverse:    reverse,
 				},
 			)
 			if err != nil {
@@ -292,7 +301,9 @@ $ %s query gov proposals --voter cosmos1skjwj5whet0lpe65qaq4rpq03hjxlwd9nf39lk
 	}
 
 	cmd.Flags().String(flagVoter, "", "(optional) filter by proposals voted on by voted")
+	cmd.Flags().Bool(flagReverse, true, "proposal ordering flag")
 	flags.AddQueryFlagsToCmd(cmd)
+	flags.AddPaginationFlagsToCmd(cmd, "customgov")
 
 	return cmd
 }