+++ x/tokens/types/expected_keepers.go
@@ -8,7 +8,7 @@ import (
 // CustomGovKeeper defines the expected interface contract the tokens module requires
 type CustomGovKeeper interface {
 	CheckIfAllowedPermission(ctx sdk.Context, addr sdk.AccAddress, permValue customgovtypes.PermValue) bool
-	GetNextProposalID(ctx sdk.Context) (uint64, error)
+	GetNextProposalIDAndIncrement(ctx sdk.Context) uint64
 	GetNetworkProperties(ctx sdk.Context) *customgovtypes.NetworkProperties
 	SaveProposal(ctx sdk.Context, proposal customgovtypes.Proposal)
 	AddToActiveProposals(ctx sdk.Context, proposal customgovtypes.Proposal)